Refactor FXIOS-4509 credential provider scripts
