+++ .github/workflows/update_credential_provider_script.yml
@@ -1,6 +1,8 @@
 name: Update credential provider script
 
-on: workflow_dispatch
+on:
+  schedule:
+    - cron: '0 0 0 * * *'
 
 jobs:
   build:
