+++ test-fixtures/requirements.txt
@@ -2,3 +2,6 @@ semver
 requests
 ruamel.yaml
 pygithub
+filecmp
+shutil
+os
